Polyfill native Node modules
Having access to native node modules is required by the upcoming
features so here we're polyfilling them.

While there are several approaches to accomplish this (most
popular of which is RN Nodeify), we decided to go with
"node-libs-react-native" as it looked a lot simpler and got the
job done. References:

- https://github.com/parshap/node-libs-react-native#node-libs-react-native
- https://github.com/tradle/rn-nodeify

Also note that the "react-native-randombytes" library is installed
explicitly as it is a peer dependency of "react-native-crypto"
(which is the substitute of Node's "crypto"), and it needs to be
linked to native modules. Reference:
https://github.com/mvayngrib/react-native-crypto#install