Move assistant icons into the assistant directory
I think we should have the assistant icons in the assistant
directory as they are rather tightly coupled to it. Not having a
central icon directory may result in some duplication, but it
might be ok as aggressive componentization / encapsulation may
prove to be more useful in the long run.