Introduce date & date range picker widgets
I'm not very happy with the current library. It seems a bit slow
and also depends on Moment which is a big dependency. But since
it's simple and does the job properly, I'm ok with it for now.
Let's keep this in mind and replace it with a better performing
library (if exists) in the future.