Refactor the claim creation scene
Move the submit logic from the root component to the claim form,
because it is there we will need to handle the backend response
and show user some feedback.

As much as I like to push "side effects" to the boundaries, in
this case this feels like the way to go.