Grand refactoring
This is not a pure refactoring but rather a mix of a refactoring
along with a set of changes that were made necessary by the recent
migration on the ixo backend.

The changes are too much, which is not really a good practice; but
several parts of the project were starting to get out of hand. I
had to quickly get the project back on the right track as well as
deal with the breaking changes caused by the migration.

I apologize for this, but sometimes these things are necessary.